refactor(activerecord): fold _performUpdate into Persistence#_updateRecord
Removes the trails-only `_performUpdate` decomposition from `base.ts` and
inlines its body in `persistence.ts#_updateRecord`, the way Rails writes
`Persistence#_update_record` (persistence.rb:900-916). The two early returns
collapse into Rails' single `if attribute_names.empty?` branch, so
`@previously_new_record = false` and the `save(&block)` yield now run on that
arm too — as they always did via the caller.

With `_performUpdate` gone, the `_pendingOperation` promise slot it parked the
UPDATE in has no remaining writer: the statement is awaited inline. The
write-path bind helper (`writePathValueNode`) moves alongside the body it
serves, and the unused `_performUpdate` member drops out of `dirty.ts`' host
interface.
